requirements.txt
fastapi
ipykernel
jamo
jiter
librosa
matplotlib
numpy
openai==1.52.2
pandas
pillow
scikit-learn
scipy
uvicorn
pydub
regex
python-multipart
praat-parselmouth
fastdtw